+++ styleguide/components/Section.js
